// a2m-afp-front/src/model/doc/AfpMenu.ts
export class AfpMenu {
    id?: string;
    lev?: string;
    name?: string;
    nameKr?: string;
    url?: string;
    menuParentId?: string;
    orderNo?: string;
    useYn?: any;
    createdDate?: any;
    createdby?: any;
    updatedDate?:any;
    updatedBy?:any;
    tsstMenuId?: any;
    children?: AfpMenu[];
    state?: any;
};